// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  admin: typeof admin;
  auth: typeof auth;
  credits: typeof credits;
  http: typeof http;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;